create countdown

+++ truffle/contracts/CardBattle.sol
@@ -7,12 +7,12 @@ contract CardBattle is CardService {
     uint8 winRate = 50;
     uint8 rareEggRate = 30;
 
-    modifier checkCooldown(uint _cardId) {
+    modifier checkCooldown(uint256 _cardId) {
         require(listCard[_cardId].readyTime <= block.timestamp, "is cooldown");
         _;
     }
 
-    function _triggerCooldown(uint _cardId) internal {
+    function _triggerCooldown(uint256 _cardId) internal {
         listCard[_cardId].readyTime = uint32(block.timestamp + cooldownTime);
     }
 
@@ -20,15 +20,21 @@ contract CardBattle is CardService {
         return (listCard[_random(listCard.length)]);
     }
 
-    function _battle(uint _cardId) external checkCooldown(_cardId) onlyOwnerOfCard(_cardId) returns (uint8, uint8) {
+    function _battle(uint256 _cardId)
+        external
+        checkCooldown(_cardId)
+        onlyOwnerOfCard(_cardId)
+        returns (uint8, uint8)
+    {
         IGameItems nft = IGameItems(gameItemAddress);
         uint8 checkWin = uint8(_random(100));
         uint8 checkEgg = uint8(_random(100));
         if (checkWin >= winRate) {
-            if (checkEgg >= rareEggRate) nft.userMintEgg(msg.sender, nft.getCommonEggId());
+            if (checkEgg >= rareEggRate)
+                nft.userMintEgg(msg.sender, nft.getCommonEggId());
             else nft.userMintEgg(msg.sender, nft.getRareEggId());
         }
         _triggerCooldown(_cardId);
-        return (checkWin, checkEgg) ;
+        return (checkWin, checkEgg);
     }
 }
