+++ truffle/contracts/GameItems.sol
@@ -5,35 +5,38 @@ import "../node_modules/@openzeppelin/contracts/token/ERC1155/ERC1155.sol";
 import "../node_modules/@openzeppelin/contracts/access/Ownable.sol";
 
 contract GameItems is ERC1155, Ownable {
-
     // egg properties
-    uint constant CommonEgg = 2 ** 10 - 2;
-    uint constant RareEgg = 2 ** 10 - 3;
+    uint256 constant CommonEgg = 2**10 - 2;
+    uint256 constant RareEgg = 2**10 - 3;
 
     constructor() ERC1155("") {}
 
-    modifier isEgg(uint _nftId) {
+    modifier isEgg(uint256 _nftId) {
         require(_nftId >= 2**10 - 3, "user cannot mint this nft");
         _;
     }
 
-    function userMintCard(address _sender, uint _id) external {
+    function userMintCard(address _sender, uint256 _id) external {
         _mint(_sender, _id, 1, "");
     }
 
-    function userMintEgg(address _sender, uint _typeEgg) external {
+    function userMintEgg(address _sender, uint256 _typeEgg) external {
         _mint(_sender, _typeEgg, 1, "");
     }
 
-    function getCommonEggId() external pure returns(uint) {
+    function getCommonEggId() external pure returns (uint256) {
         return CommonEgg;
     }
 
-    function getRareEggId() external pure returns(uint) {
+    function getRareEggId() external pure returns (uint256) {
         return RareEgg;
     }
-    
-    function userBurnEgg(address _sender, uint _typeEgg, uint _amount) external {
+
+    function userBurnEgg(
+        address _sender,
+        uint256 _typeEgg,
+        uint256 _amount
+    ) external {
         _burn(_sender, _typeEgg, _amount);
     }
 }
